backend/requirements.txt
fastapi
uvicorn
kafka-python
requests
python-dotenv